Fix rendering garbage and screen centering for most games
Fixing the garbage on the left and right edge of the screen.
Fixing screen centering for most games. Mortal Kombat now looks correct.
For some reason the screen dimensions reported to osd_create_display()
are different then the actual bitmap dimensions which was throwing off
the screen centering. Mr Do! still is not centered horizontally though.
Tested a lot of games and most of them seem to work. NBA Jam does not
crash now but it still does not work properly nor does some of the
newer Mortal Kombat games.